pyproject.toml
[project]
name = "agentic-ai-hr-automation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "anthropic>=0.76.0",
    "beautifulsoup4>=4.14.3",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "langchain>=1.2.4",
    "langchain-anthropic>=1.3.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "llama-cloud>=0.1.45",
    "llama-cloud-services>=0.6.90",
    "llama-index-workflows>=2.12.2",
    "ollama>=0.6.1",
    "openai>=2.15.0",
    "protobuf>=5.29.5",
    "pydantic>=2.12.5",
    "pymongo[srv]>=4.16.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "python-ulid>=3.1.0",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.40.0",
    # New features dependencies (v1.1.0)
    "xlsxwriter>=3.0.0",  # Excel export functionality
    # MinIO Storage dependencies (v1.2.0)
    "minio>=7.0.0",  # MinIO object storage client
    # Authentication and Authorization dependencies (v1.3.0)
    "python-jose[cryptography]>=3.3.0",  # JWT token creation and validation
    "passlib>=1.7.4",  # Password hashing (removed [bcrypt] extra)
    "bcrypt>=4.0.0",  # Bcrypt password hashing algorithm
    "python-multipart>=0.0.21",  # Form data parsing
    "sqlalchemy>=2.0.0",  # SQL toolkit
    "psycopg2-binary>=2.9.9",  # PostgreSQL adapter
    "alembic>=1.13.0",  # Database migration
    "email-validator>=2.3.0",  # Email validation
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend*"]
exclude = ["frontend*", "venv*", "bin*", "docs*", "tests*"]
